requirements.txt
numpy==1.19.5
matplotlib==2.2.3
future==0.18.2
mysql-connector-python==8.0.29